Fix the build on i686

+++ bfffs-core/tests/functional/fs.rs
@@ -3221,7 +3221,8 @@ root:
         let mtime = Timespec {sec: 3, nsec: 4};
         let ctime = Timespec {sec: 5, nsec: 6};
         let birthtime = Timespec {sec: 7, nsec: 8};
-        let flags = libc::UF_NODUMP;
+        #[allow(clippy::useless_conversion)]    // Not useless on i686
+        let flags = u64::from(libc::UF_NODUMP);
 
         let assert = |attr: GetAttr| {
             assert_eq!(attr.nlink, 1); // Shouldn't have been changed
